requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
matplotlib==3.9.2
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
supabase==2.11.0
pydantic==2.10.4
